+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 3.10.3
 
+- Improvement: Updated module icon
 - Improvement: `ConfigureValidationRules` method is no longer be created on _validators_ if it would be created empty.
 
 ### Version 3.10.2